% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_feedback.R
\name{all_feedback}
\alias{all_feedback}
\title{Generate All Feedback (Glasgow-specific)}
\usage{
all_feedback(markfile, id_col = 1, moodlefile = NA, evalfile = NA,
  template_file = FALSE, moodle_dir = FALSE)
}
\arguments{
\item{markfile}{path to marking file or directory of marking files}

\item{id_col}{column where the participant id number is found (can be named or numeric) (defaults to 1)}

\item{moodlefile}{path to file of student_id and moodle_id matches}

\item{evalfile}{path to evaluation criteria file (optional)}

\item{template_file}{path to the .Rmd template file for feedback (default of FALSE uses template called {class_id}_template.Rmd for each class}

\item{moodle_dir}{give the directory a moodle-style name, use the pattern 'Participant_*_assignsubmission_onlinetext_' where \* is replaced by the moodle ID (defaults FALSE)}
}
\value{
list
}
\description{
\code{all_feedback} generates feedback for all files in a directory
}
\examples{
load_marks("001_marking.csv", "001_submissions", "001_eval.csv")
}
