% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_dist.R
\name{cat_dist}
\alias{cat_dist}
\title{Category Distribution Graph}
\usage{
cat_dist(marking, rating_levels = 1:3)
}
\arguments{
\item{marking}{list containing a dataframe (marks)}

\item{rating_levels}{list of rating levels (default 1:3)}
}
\value{
ggplot
}
\description{
\code{cat_dist} create a graph of the category distribution
}
\examples{
cat_dist(marking_example, rating_levels = 1:4)
}
