% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_mark.R
\name{cat_mark}
\alias{cat_mark}
\title{Category vs Mark Graph}
\usage{
cat_mark(marking)
}
\arguments{
\item{marking}{list containing a dataframe (marks) and eval}
}
\value{
ggplot
}
\description{
\code{cat_mark} create a graph of the category sums versus the marks
}
\examples{
cat_mark(marking_example)
}
