% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_table.R
\name{category_table}
\alias{category_table}
\title{Format Category Table}
\usage{
category_table(ind_fb, eval, cats)
}
\arguments{
\item{ind_fb}{1-row data frame of assignment evaluations}

\item{eval}{data frame of evaluation labels (q) and descriptions (Category)}

\item{cats}{list of labels to map onto evaluations}
}
\value{
A latex-formatted table
}
\description{
\code{category_table} returns a latex-formatted table
from an existing 1-row dataframe (ind_fb) with specified 
evaluation categories (eval) and category lables (cats).
}
\examples{
marking_example$marks \%>\% 
  slice(1) \%>\% 
  category_table(
    marking_example$eval, 
    c("Excellent", "Good", "Poor")
  )
}
