% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_second_marking.R
\name{check_second_marking}
\alias{check_second_marking}
\title{Create Second Marking Table}
\usage{
check_second_marking(marking, filename = NA)
}
\arguments{
\item{marking}{list containing a dataframe (marks) and moodle participant directory (dir)}

\item{filename}{path of second marking file}
}
\value{
none
}
\description{
\code{check_second_marking} returns a table with first and second marks
}
\examples{
check_second_marking(marking_example)
}
