% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_grades.R
\name{convert_grades}
\alias{convert_grades}
\title{Convert Grades to Marks}
\usage{
convert_grades(grades, scale = glasgow22())
}
\arguments{
\item{grades}{vector of letter or number grades}

\item{scale}{conversion table containing `letters` and `numbers`}
}
\value{
vector
}
\description{
\code{convert_grades} convert letter grades to numeric marks
}
\examples{
convert_grades(marking_example$marks$Grade)
}
