% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_grades.R
\name{glasgow22}
\alias{glasgow22}
\title{Glasgow 22-Point Marking Scale}
\usage{
glasgow22()
}
\value{
data frame
}
\description{
\code{glasgow22} The University of Glasgow 22-point marking scale conversion table
}
\examples{
glasgow22()
}
