% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_marks.R
\name{load_marks}
\alias{load_marks}
\title{Load Marking Data}
\usage{
load_marks(markfile = "marks.csv", dir = NA, evalfile = NA,
  assign_id = NA, id_col = 1, encoding = "latin1")
}
\arguments{
\item{markfile}{path to marking file or directory of marking files}

\item{dir}{path to moodle submission directory (optional)}

\item{evalfile}{path to evaluation criteria file (optional)}

\item{assign_id}{Moodle assignment ID}

\item{id_col}{column where the participant id number is found (can be named or numeric) (defaults to 1)}

\item{encoding}{encoding for importing .csv files (defaults to latin1)}
}
\value{
list
}
\description{
\code{load_marks} loads marking data
}
\examples{
load_marks("001_marking.csv", "001_submissions", "001_eval.csv")
}
