% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_feedback.R
\name{make_feedback}
\alias{make_feedback}
\title{Create Feedback Reports}
\usage{
make_feedback(marking, template, filename = "feedback.pdf",
  moodle_dir = FALSE, quiet = TRUE, filetype = "pdf")
}
\arguments{
\item{marking}{list containing a dataframe (marks) and moodle participant directory (dir)}

\item{template}{path to the .Rmd template file for feedback}

\item{filename}{the name to give the rendered file (defaults to feedback.pdf)}

\item{moodle_dir}{give the directory a moodle-style name, use the pattern 'Participant_*_assignsubmission_onlinetext_' where \* is replaced by the moodle ID (defaults FALSE)}

\item{quiet}{passed to rmarkdown::render (default TRUE)}

\item{filetype}{output a pdf or html file (defaults to pdf)}
}
\value{
none
}
\description{
\code{make_feedback} creates feedback PDFs from a template .Rmd file
and saves them in the moodle directory under the appropriate
participant folder
}
\examples{
make_feedback(marking_example, "787420_template.Rmd")
}
