% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report.R
\name{make_report}
\alias{make_report}
\title{Create Assessment Summary Report}
\usage{
make_report(marking, template = "report_template.Rmd", show = c("eval",
  "mark2", "mark_dist", "cat_dist", "cat_mark"), output_file = NA,
  quiet = TRUE)
}
\arguments{
\item{marking}{list containing a dataframe (marks)}

\item{template}{path to the .Rmd template file for reports}

\item{show}{list of sections to show in the default report template}

\item{output_file}{path for the rendered file (defaults to feedback.pdf)}

\item{quiet}{passed to rmarkdown::render (default TRUE)}
}
\value{
none
}
\description{
\code{make_feedback} creates a report PDF from a dataframe of 
assignment evaluations
}
\examples{
make_report(marking_example)
}
