\name{markr-package}
\alias{markr-package}
\alias{markr}
\docType{package}
\title{
\packageTitle{markr}
}
\description{
\packageDescription{markr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{markr}
\packageIndices{markr}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~

markr provides a visual interface for rapid anomaly detection in time series data. 
It can be used as part of an analysis workflow or stand alone, with examples included. 
The package fits a machine learning algorithm to automatically suggest anomalies. 
The user can keep those flags or mark new sections of the data and add flag names and comments. 
Marked data and flag metadata can be saved for future use. 
The fitted model objects can be saved locally for training future data sets as well.
}
\author{
\packageAuthor{Aaron Berdanier}

Maintainer: \packageMaintainer{Aaron Berdanier, aaron.berdanier@gmail.com}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
