% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markr_app.R
\name{markr_app}
\alias{markr_app}
\title{Launch Shiny App}
\usage{
markr_app(...)
}
\arguments{
\item{...}{arguments to pass to shiny::runApp}
}
\description{
The app will eventually be a fully interactive way to create marking feedback from spreadsheets. It currently only lets you create YAML files interactively.
}
\examples{
\dontrun{ markr_app() }

}
