% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\docType{data}
\name{markr_env}
\alias{markr_env}
\title{Store user information}
\format{An object of class \code{environment} of length 0.}
\usage{
markr_env
}
\arguments{
\item{str}{the ID that the user logged into the tutorial with}
}
\description{
These are internal functions
}
\keyword{datasets}
