% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mark.R
\name{predict_mark}
\alias{predict_mark}
\title{Linear Model of Marks}
\usage{
predict_mark(marking)
}
\arguments{
\item{marking}{list containing a dataframe (marks)}
}
\value{
lm() summary
}
\description{
\code{predict_mark} generate a linear model of marks from category ratings
}
\examples{
predict_mark(marking_example)
}
