% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_marks.R
\name{read_marks}
\alias{read_marks}
\title{Read marks from spreadsheet and/or yaml}
\usage{
read_marks(tbl = NULL, yaml = NULL, join_by = NULL, ...)
}
\arguments{
\item{tbl}{path of tabular data (e.g., excel, csv)}

\item{yaml}{path of yaml data}

\item{join_by}{column name to join by if reading both tabular and yaml}

\item{...}{Extra column values to be added}
}
\value{
data table
}
\description{
Read marks from spreadsheet and/or yaml
}
\examples{
tbl <- system.file("basic", "marks.csv", package = "markr")
yaml <- system.file("basic", "fb.yml", package = "markr")
marks <- read_marks(tbl, yaml, join_by = "ID")
marks101 <- read_marks(tbl, yaml, join_by = "ID", class = "PSYCH101")
}
