% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_marks.R
\name{save_marks}
\alias{save_marks}
\title{Save full marking list}
\usage{
save_marks(marking, filename = NA)
}
\arguments{
\item{marking}{list containing a dataframe (marks)}

\item{filename}{path to save}
}
\value{
none
}
\description{
\code{save_marks} save the amalgamated marking list
}
\examples{
save_marks(marking, "full_marks.xls")
}
