% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_mark.R
\name{second_mark}
\alias{second_mark}
\title{Generate Second Marking List}
\usage{
second_mark(marking, dir = "second_marking",
  remove_file = "feedback.pdf", pass_min = 9,
  show_first_mark = FALSE)
}
\arguments{
\item{marking}{list containing a dataframe (marks) and moodle participant directory (dir)}

\item{dir}{directory name to save second marking in}

\item{remove_file}{filename to remove from second marking directory}

\item{pass_min}{minimum passing mark; all lower marks are second marked (default 9)}

\item{show_first_mark}{include first mark in second marking files (default FALSE)}
}
\value{
none
}
\description{
\code{second_mark} determines what assessments need to be second marked 
and bundles files into a folder with the second marking sheet
}
\examples{
second_mark(marking, "data-raw/example/submissions")
}
