% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_marks.R
\name{update_marks}
\alias{update_marks}
\title{Update Marks}
\usage{
update_marks(marking, filename)
}
\arguments{
\item{marking}{list containing a dataframe (marks) and moodle participant directory (dir)}

\item{filename}{path to updated marking file (required columns: moodle_id, question, updated_grade)}
}
\value{
list
}
\description{
\code{update_marks} returns a table with first and second marks
}
\examples{
update_marks(marking_example, "final_marks.csv")
}
