% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_table.R
\name{category_table}
\alias{category_table}
\title{Category Table}
\usage{
category_table(
  marks,
  cols,
  cats = NULL,
  symbol = "&#10004;",
  kable = TRUE,
  critwidth = 0.5,
  ...
)
}
\arguments{
\item{marks}{table with criteria columns (only the first row will be used)}

\item{cols}{a named vector of the criteria (names are the marks column names and values are the text to display)}

\item{cats}{a named vector of the marking categories (names are the value in marks and values are the text to display)}

\item{symbol}{the symbol to display in the table (✔ is the HTML code for a check mark)}

\item{kable}{return kable (T) or data table (F)}

\item{critwidth}{the width of the first (criteria) column; the rest of the columns are set to equal widths}

\item{...}{arguments to pass to kableExtra}
}
\value{
a kable table or a data table
}
\description{
Category Table
}
\examples{
marks <- data.frame(ID = 7, K = 1, E = 2, C = 3)
cols <- c(K = "Knowledge", E = "Evaluation", C = "Communication")
cats <- c("1" = "Bad", "2" = "OK", "3" = "Good")
category_table(marks, cols, cats) # html table
category_table(marks, cols, cats, "X", FALSE) # data table

}
