% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_grades.R
\name{convert_grades}
\alias{convert_grades}
\title{Convert Grades to Marks}
\usage{
convert_grades(grades, to = NA, scale = glasgow22())
}
\arguments{
\item{grades}{vector of letter or number grades}

\item{to}{convert to \code{letters} or \code{numbers} (default NA converts to other type)}

\item{scale}{conversion table containing \code{letters} and \code{numbers}}
}
\value{
vector
}
\description{
\code{convert_grades} convert letter grades to numeric marks
}
\examples{
convert_grades("A1") # convert from letter to number
convert_grades(1:22) # convert from number to letter

# make all letter grades the same case
convert_grades(c("a1", "B2", "c3"), "letters")
}
