% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_feedback.R
\name{make_feedback}
\alias{make_feedback}
\title{Create Feedback Reports}
\usage{
make_feedback(marks, template, filename = NULL, group_by = NULL, quiet = FALSE)
}
\arguments{
\item{marks}{data frame of marks or path to file}

\item{template}{path to the .Rmd template file for feedback}

\item{filename}{path to save files, defaults to "feedback/[group_by columns]", but you can reference any column names inside square brackets (e.g., "assignment_[moodle_id]_/[Student ID]")}

\item{group_by}{columns to group by (e.g., if the marking file contains multiple rows per student) defaults to grouping by row if NULL}

\item{quiet}{print a message for each rendering}
}
\description{
\code{make_feedback} creates feedback files from a template .Rmd file
}
\examples{
\dontrun{
# saves each feedback file in a separate folder with the ID
make_feedback(demo_marks, "template.Rmd", "fb/[ID]/exam_Q[question]", "ID")
}
}
