% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_dist.R
\name{mark_dist}
\alias{mark_dist}
\title{Marking Distribution Graph}
\usage{
mark_dist(
  marks,
  mark_col,
  fill_col = mark_col,
  facet_by = NULL,
  scale = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{marks}{data frame of marking info}

\item{mark_col}{column where the mark is found}

\item{fill_col}{column for setting fill}

\item{facet_by}{columns for facet graphs}

\item{scale}{vector of values for the x-axis (full range of the marking scale), defaults to alphabetical or numeric ordering of existing data}

\item{na.rm}{remove NA values in the mark_col}
}
\value{
ggplot
}
\description{
\code{mark_dist} create a graph of the marking distribution
}
\examples{
mark_dist(demo_marks, "mark", scale = 0:5)
}
