% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markr_example.R
\name{markr_example}
\alias{markr_example}
\title{Make example directory}
\usage{
markr_example(dir = "markr_example")
}
\arguments{
\item{dir}{the directory to save the example files in}
}
\value{
opens the demo.Rmd file for editing
}
\description{
Make example directory
}
\examples{
\dontrun{
markr_example("demo")
}
}
