% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2yaml.R
\name{tbl2yaml}
\alias{tbl2yaml}
\title{Convert table to YAML}
\usage{
tbl2yaml(tbl = NULL, filename = NULL, ..., open = TRUE)
}
\arguments{
\item{tbl}{The data table or path to a data table}

\item{filename}{The name of a file to save to}

\item{...}{Extra column values to be added}

\item{open}{Whether to open the file in a text editor}
}
\value{
opens file or returns YAML text
}
\description{
Convert table to YAML
}
\examples{
tbl <- system.file("example", "marks.csv", package = "markr")
yaml <- tbl2yaml(tbl, class = "PSYCH1001")

}
