% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password.R
\name{user_ID}
\alias{user_ID}
\title{Login ID/Password question}
\usage{
user_ID(
  email = NULL,
  key = NULL,
  store = NULL,
  placeholder = "Enter ID here in format user_name::password"
)
}
\arguments{
\item{email}{email address of the account holding the key and store sheets}

\item{key}{Data frame of user  IDs and  passwords}

\item{store}{ID of sheet storing submissions}

\item{placeholder}{Character string giving instructions to user.}
}
\description{
Creates a text box group tutorial quiz question.
}
\examples{
\dontrun{
user_ID("03829392912232232")
}
}
